+++ CHANGES.md
@@ -1,4 +1,5 @@
 #### [unreleased]
+* don't load `pluggable.php` too early, call `wp_create_nonce()` in `plugins_loaded` hook
 
 #### 10.6.0 / 2021-09-23
 * loads of security updates, nonce all the things